# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Gestion de Incidentes API",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/estadisticas/tendencias": {
      "get": {
        "tags": [
          "Estadisticas"
        ],
        "summary": "Obtener tendencias de incidentes agrupadas por dia o mes",
        "description": "Retorna una serie temporal con la cantidad de incidentes por periodo\n(dia o mes) dentro del rango de fechas especificado.\n\nIncluye desglose por sector y distribucion de estados para el rango completo.",
        "operationId": "get_tendencias_api_v1_estadisticas_tendencias_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "agrupar_por",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/AgruparPorEnum",
              "description": "Granularidad de agrupacion: 'dia' o 'mes'"
            },
            "description": "Granularidad de agrupacion: 'dia' o 'mes'"
          },
          {
            "name": "desde",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date",
              "description": "Fecha de inicio del rango (YYYY-MM-DD)",
              "title": "Desde"
            },
            "description": "Fecha de inicio del rango (YYYY-MM-DD)"
          },
          {
            "name": "hasta",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date",
              "description": "Fecha de fin del rango (YYYY-MM-DD)",
              "title": "Hasta"
            },
            "description": "Fecha de fin del rango (YYYY-MM-DD)"
          },
          {
            "name": "sector_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filtrar por sector responsable (opcional)",
              "title": "Sector Id"
            },
            "description": "Filtrar por sector responsable (opcional)"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TendenciasResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/estadisticas/resumen": {
      "get": {
        "tags": [
          "Estadisticas"
        ],
        "summary": "Obtener resumen de KPIs del dashboard",
        "description": "Retorna KPIs agregados: total de incidentes, promedio diario,\ndistribucion por sector y estado, y tasa de revision humana.\n\nSi no se especifican fechas, se usa el rango de los ultimos 30 dias.",
        "operationId": "get_resumen_api_v1_estadisticas_resumen_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "desde",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Fecha de inicio del rango. Default: 30 dias atras.",
              "title": "Desde"
            },
            "description": "Fecha de inicio del rango. Default: 30 dias atras."
          },
          {
            "name": "hasta",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Fecha de fin del rango. Default: hoy.",
              "title": "Hasta"
            },
            "description": "Fecha de fin del rango. Default: hoy."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResumenResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "AgruparPorEnum": {
        "type": "string",
        "enum": [
          "dia",
          "mes"
        ],
        "title": "AgruparPorEnum",
        "description": "Granularidad de agrupacion temporal para la serie de tendencias."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PeriodoInfo": {
        "properties": {
          "desde": {
            "type": "string",
            "title": "Desde",
            "description": "Fecha de inicio del rango consultado"
          },
          "hasta": {
            "type": "string",
            "title": "Hasta",
            "description": "Fecha de fin del rango consultado"
          },
          "agrupar_por": {
            "type": "string",
            "title": "Agrupar Por",
            "description": "Granularidad usada: 'dia' o 'mes'"
          }
        },
        "type": "object",
        "required": [
          "desde",
          "hasta",
          "agrupar_por"
        ],
        "title": "PeriodoInfo",
        "description": "Metadatos del rango consultado en la respuesta de tendencias."
      },
...
      "ResumenResponse": {
        "properties": {
          "total_incidentes": {
            "type": "integer",
            "title": "Total Incidentes",
            "description": "Total de incidentes en el rango"
          },
          "promedio_diario": {
            "type": "number",
            "title": "Promedio Diario",
            "description": "Promedio de incidentes por dia en el rango"
          },
          "distribucion_sectores": {
            "additionalProperties": {
              "type": "integer"
            },
            "type": "object",
            "title": "Distribucion Sectores",
            "description": "Distribucion de incidentes por sector"
          },
          "distribucion_estados": {
            "additionalProperties": {
              "type": "integer"
            },
            "type": "object",
            "title": "Distribucion Estados",
            "description": "Distribucion de incidentes por estado"
          },
          "tasa_revision_humana": {
            "type": "number",
            "title": "Tasa Revision Humana",
            "description": "Proporcion de incidentes que requirieron revision humana (0.0 a 1.0)"
          }
        },
        "type": "object",
        "required": [
          "total_incidentes",
          "promedio_diario",
          "tasa_revision_humana"
        ],
        "title": "ResumenResponse",
        "description": "Respuesta del endpoint GET /api/v1/estadisticas/resumen."
      },
...
      "SerieTemporal": {
        "properties": {
          "periodo": {
            "type": "string",
            "title": "Periodo",
            "description": "Etiqueta del periodo: YYYY-MM-DD (dia) o YYYY-MM (mes)"
          },
          "total": {
            "type": "integer",
            "title": "Total",
            "description": "Cantidad de incidentes en este periodo"
          },
          "por_sector": {
            "additionalProperties": {
              "type": "integer"
            },
            "type": "object",
            "title": "Por Sector",
            "description": "Desglose de incidentes por nombre de sector (ej. {'Sistemas': 10, ...})"
          }
        },
        "type": "object",
        "required": [
          "periodo",
          "total"
        ],
        "title": "SerieTemporal",
        "description": "Un punto en la serie temporal de tendencias (un periodo)."
      },
      "TendenciasResponse": {
        "properties": {
          "periodo": {
            "$ref": "#/components/schemas/PeriodoInfo"
          },
          "total_incidentes": {
            "type": "integer",
            "title": "Total Incidentes",
            "description": "Total de incidentes en el rango completo"
          },
          "series": {
            "items": {
              "$ref": "#/components/schemas/SerieTemporal"
            },
            "type": "array",
            "title": "Series"
          },
          "distribucion_sectores": {
            "additionalProperties": {
              "type": "integer"
            },
            "type": "object",
            "title": "Distribucion Sectores",
            "description": "Distribucion agregada de incidentes por sector en el rango"
          },
          "distribucion_estados": {
            "additionalProperties": {
              "type": "integer"
            },
            "type": "object",
            "title": "Distribucion Estados",
            "description": "Distribucion agregada de incidentes por estado en el rango"
          }
        },
        "type": "object",
        "required": [
          "periodo",
          "total_incidentes"
        ],
        "title": "TendenciasResponse",
        "description": "Respuesta del endpoint GET /api/v1/estadisticas/tendencias."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/v1/auth/login"
          }
        }
      }
    }
  }
}